Fixes for "warning: subscript has type `char'" The problem is that isspace and isdigit are macros under solaris

+++ ptclassic/src/octtools/Packages/octmm/hackterm.c
@@ -269,7 +269,7 @@ char *match;
     while (*match) {
 	if (*termName++ != *match++) return(OCTMM_TRUE);
     }
-    while (isdigit(*termName)) {
+    while (isdigit((int)*termName)) {
 	termName++;
     }
     match = SUFFIX_CLOSE;
